WIP: convert GC Finalizers to Close() methods.
After discussions with @kalbasit around issue #11, we concluded it was
probably better for the developer to be freeing resources explicitly.
See also:

    https://gist.github.com/dwbuiten/c9865c4afb38f482702e#cleaning

I'm not totally sure about that anymore, in particular because it rules
out constructions like:

    db.NewQuery("foo").Threads()

Which covers the most common case.

Everything is done with the exception of fixing uses of the above; there
are still a few in the test suite.

To be specific, this patch:

* Changes DB.Close() to call notmuch_database_destroy instead of _close.
  Since we won't be doing any GC, we ought to call _destroy so the
  memory is released, not just file descriptors and locks and such.
* Remove all calls to runtime.SetFinalizer()
* Add a Close() method for each corresponding _destroy function.
* Call Close() methods in appropriate places in the test suite.

Per a comment added to the example, adding calls to Close() isn't
necessary there, since everything is live until the program exits (The
database is already being closed).

Beware: This branch will likely be destructively updated a few times
before it is finally merged.