Tweak the way iterators work.
The old API resulting in an implementation problem, which I *suspect* is
the cause of a segfault I was seeing in another program.

In particular, the Next() methods previously *copied* the contents of
the struct, which does not move the finalizer. So, when the original is
garbage collected, the finalizer runs even though there is another
reference to the underlying C object.